Switch playFile to playRawStream
I think this might fix the problem with the bot never actually playing
anything. I think that the bot is trying to load in the entire mp3, which
obviously can't happen because Cadence is a streaming server, so it looks
exactly like the bot does nothing after joining the voice channel.

According to the docs, playRawStream might fix this issue.