Add logging to end listener
I don't really care for having this logged so simply, but in the long term I intend to wrap the logger so that all log statements have a time in a better format attached, so this is basically a placeholder.

This statement, however, needs to log the time, so I can compare with @kenellorando to see if something happened on Cadence's end.