Reuse nowplaying command to log nowplaying on stream end
This also refactors the message handler into its own function, so that
this may be done.

Also, I'm rather pleased with this solution. This should accurately
(enough) mock the Discord message, so nowplaying can't tell its not
interacting with an actual user.

I guess we'll see in testing whether it actually works.