Change stream end message to a retrying message
In the next few commits, I will be changing the behavior of the bot from
stopping playback when an error occurs to reconnecting automatically to
Cadence when an error occurs. This will reduce the manual oversight
required to keep CadenceBot running on a server long-term, especially if
there's a desire for the bot to run uninterrupted.
This commit is the first part of that, changing the bot's interaction with
the user to match upcoming behavior.

The old behavior was valuable in that it kept a record of the reported
error. Therefore, since this commit removes that message, it adds a log
entry containing the error message.