Change stream end behavior to reconnect
This is done by sending a mocked message to command(), containing the same
content, same reply function, and the same member voice channel, as the
original play message did at the time it was received.

Why is it done this way instead of by simply recalling command() with the
original message? Here's a copy of the explanation I gave in a comment in
this commit:

...we want to reissue ourselves a play command equivalent to the original
one, to begin playback on the same channel.
At a glance, that means reissuing the original message.
However, if the user has since disconnected... We'll generate a spurious
error.
The play code wants to connect to the user's channel: It doesn't know what
channel to connect to if the user isn't connected.
We, however, do.
So, if there isn't a [voice channel], we need to mock it.
At the same time, the user could be in the wrong [voice channel]. That would make us connect to the incorrect channel.
Basically, we just generally want to mock the [voice channel].
That's why the naÃ¯veimplementation (command(message)) isn't the one we
use here.