Add auth.json creation to first-time-setup.sh
The script should now be able to take a freshly cloned instance of
CadenceBot and make it ready to be restarted, with minimal user input.
This is very good, as that means that the only step the user must remember
is how to fetch their bot token from discord (and there's a link in the
README for that)

I'll leave auth-example.json in place for users that prefer to do this
manually.