Switch info and notice colors
Originally, info was going to be the basis. I soon realized that I wanted to have a level between baseline and debug (aka not normally logged), so I upgraded the basis to notice. Therefore, I'm switching their configured colors so that the log will continue to have 'normal' level logging in white.