Correct infinite reconnect attempts bug
I realized that the bug was caused by the falsified play command overwriting the reconnectAllowedAt set in the recovery code.

I added the reconnectAllowedAt setup in play to prevent undefined lookups later, while reworking data storage to prevent interference between channels. However, I didn't realize, because I didn't sufficiently think about the interaction of commands with this sort of thing being done, that this would overwrite the same data structure being assigned earlier during a reconnect. This effectively disabled the timeout (setting it to 0 seconds), and thus allowed the classic infinite reconnect behavior to return (requiring @kenellorando to temporarily kick the bot - oops).

This fix changes behavior very slightly - Before, the timeout functioned as "a second reconnect attempt within 30 seconds is prevented", but now the timeout behavior is "a reconnect attempt within 30 seconds of starting playback is prevented".

Even though this commit is a minor change, it fixes a bug serious enough that I'll declare it a new patch numbered version.