Implement one-step request for one-result searches
That is, any string which is not parsed as a number, which only gives one result when searched in Cadence, results in a request for that result. I call this the trivial case for one-step requests.

Currently, this is supposed to preserve the state of lastSearchedSongs (as this being overwritten is a symptom of the implementation, but not intuitive behavior). I intend to take a look at this shortly.