Fix lastSearchedSongs persistence
As it turns out, the problem was that the callback for request would only run after lastSearchedSongs was reset. This meant that the request acknowledgement code would clear lastSearchedSongs immediately *after* we set it.

To fix this, we set it to reset lastSearchedSongs one second after the end of the command() call returns. Presumably, one second will be enough time for Cadence to send its reply to the bot, and thus properly allow lastSearchedSongs to be reset after it is cleared.

This would be... cleaner, to say the least, if we used a more elegant technique to issue the one-step request, as compared to the mocked request command. However, this would mean either duplication of the request code (as most of the boilerplate code would be the same, except for one change to the callback for post-request) or a significant refactor of at least the ARIA integration code. I find this impracticable at the moment, considering I intended to be in bed 16 minutes ago.