Create global for one-step-request filters
Now that we know the basic principle of one-step-request works (with the
trivial case), its time to expand the number of searches for which
one-step algorithm can select a song to request

This code will be implemented using an extensible set of filters, rather
than hardcoding the checks. This is less efficient (as it would be
possible to optimize several passes for title-match, artist-match,
title-artist-match, artist-title-match,et cetera into a single pass doing
many checks, but not possible to compress several filter function calls
into one pass. The algorithm will still be O(n), but with a larger
coefficient (more properly, it's O(n*m), where n is the number of songs
and m is the number of filters). We do, however, expect the number of
search results and the number of filters to be low, so this isn't
important.

What this achieves is making it easy to add or remove filters, in order,
and separating the detection code from the request code.

In theory, it would be possible to use this filter code to implement an
ANN which learns by watching searches and requests and acts as the last
filter before manual search to predict the best result.

In practice, this will just be a convenient way to run several checks in
sequence without cluttering command() any more than I already have.

The filters I plan to implement in this version are:

 - Single-response (trivial case): If the search results have length 1,
   request the returned song

 - Matching title (title case): If exactly one result has a title which
   exactly matches (case insensitive)  the requested string, request that
   song

 - Song by artist pattern (song+artist case): If exactly one result has
   the request string start with its song title, and that result has the
   request string end with its artist (both case insensitive), request
   that song

 - Artist/song pattern (artist+song case): If exactly one result has the
   request string start with its artist, and that result has the request
   string end with its song title (both case insensitive), request that
   song

After these, manual selection will be performed (this will be hardcoded to
ensure there exists a fallback).

While this commit includes documentation in comments regarding narrowing
filters, these are not currently supported due to lack of need. I'd like
to collect some data regarding usage to see how useful narrowing filters
would be.

My idea would be an artist-narrowing filter, removing all songs which do
not have an artist exactly matching the request. Of course, this would
have to run after all existing filters, so it would only simplify the
manual selection iff it was correct, and Cadence doesn't ever return
enough results that this would be necessary.

Therefore, the addition of such filters will be delayed until usage 'in
the wild' reveals them to be worthwhile (for example if more songs are
added to Cadence, thus cluttering search results).

Therefore, the implementation of support for these filters will be delayed
until there's some use for such a filter.