Implement title filter
This required a fix to an oversight of mine; Namely, that I forgot filters
would need the request string to do comparisons. This has been fixed, both
in the comment specifying the semantics of filters, and in the code which
calls them.

Also, note that the title filter always examines each song (unless it
finds two results). This is because my specification for the filter was
that it only select unique results - Preventing the filter from only
choosing the first of several complaint results. I am not aware of any
case where this occurs, but I would like to have the filter behave
properly from the get-go, just in case.