Add blocks of manual filter for no results
@JakobFrank found a bug where, since search doesn't edit lastSearchedSongs
for searches with no results, searching a non-term could trigger the
normal manual filter response with a message asking the user to request a
number from "Cadence returned no results".

This commit should prevent that from occurring.

Also, since, as of this commit, the one-step-request functionality is
fully operational, I hereby declare that as of this commit, we have fixed #11.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>