Update version number
Version 1.4.0 brings in the new One-Step-Request (OSR) feature, which
allows the user to instruct the bot to attempt to detect which song they'd
like to request by which string they search for.

Additionally, on the administrative side, it switches configuration to a
default+override style, allowing changes to the default configuration to
be made in version control, without interfering in personal preferences in
bot configuration.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>