Update help message for request
Removing the passage specifying 'by number' allows this message to be
inclusive of OSR, and also helps alleviate that line's problem with being
too long for its own good.

This doesn't specify that you're allowed to use a string to request, but
my experience with users shows that new users attempt to request by song
name first, before they get used to the two-step search-and-request. This
is, actually, the biggest part of why I implemented one-step-request.

So, now, its mostly a matter of teaching the very small existing community
of users that they're allowed to perform requests this way.