Allow now-playing status to be disabled in config
Because this feature is constantly polling the Cadence server for now-playing information, it can be considered a drain on the server running the bot. While the load is miniscule (on my development AWS t2.micro instance, the added CPU load is about 0.05%, plus about 7,000 bytes of network downloading and 5,000 bytes uploading - Apparently per-minute), especially in comparision to the added load introduced by the bot playing in one Discord channel (again on my instance, about 3% CPU load, plus 859,000 bytes network download and 915,000 bytes upload), it's possible that someone wants to run CadenceBot on a very limited machine, under strongly limited datacaps, or under a strongly metered connection.

That is, it's possible in theory that someone would serve CadenceBot from a machine which would overheat from the small incremental load (that is, the incremental load would cause the machine to increase in heat, even at a slow rate - Implying that the machine remains at a steady temperature without running this status updating and the increment would prevent stabilzation of heating), or over a connection which would be over its datacap with the additional half-gigabyte (combined upload and download) over a month, or over a connection which could not withstand the small additional load (one which has exactly enough bandwidth for normal use without this polling).

Because I don't want to cause anyone to have to choose to pay money to upgrade their machine or to pay for a better connection in order to continue to run CadenceBot or to shut down their copy, especially not over the added cost of an additional feature, I'm allowing it to easily be disabled, as this feature is not integral to the proper function of CadenceBot (it's basically a decoration, like sprinkles on an ice cream cone).

Of course, one can keep the feature, and lower its already-minimal requirements, by overriding config (by editing config.json) to have a longer interval (lessening the number of now playing updates).