Reduce default status interval
Since the cost of running this is so low, this shouldn't impose any real cost on my servers (since I intend to leave this feature disabled on dev most or all of the time after this goes to master.

However, it will reduce the chance of a mismatch between the status and the actually playing song at any given moment.