Update package version number
This minor version update adds the now-playing status message, which, at a
configurable interval, polls the Cadence Radio server for information on
the currently playing song, and sets it as the bot's status message. This
feature was suggested by @kenellorando.

This feature, if desired, can be disabled entirely, by setting the
interval to a negative number in `config.json`. This will set the
nowplaying information to 'Cadence Radio'.

Additionally, this version includes an update to the help message for
request, to be more inclusive of one-step-request.

Signed-off-by: Ryan Hodin <ryan951357@gmail.com>