Add logging messages to nowplaying updates
Most of this logging is set at the debug level, because there's not normally a need to read very much information. And, because of how often this function is called in normal usage, any logging at all will be pretty spammy in the log.

Therefore, at any level besides debug, only one line will appear per update (when the updates are enabled). This is a good balance between being spammy and logging information about the operation (since now playing will log relevant information at other levels). Since the function is so simple, it shouldn't be hard to track down any issues that crop up with minimal logging.