Add branch policies to README
I've been loosely complying with these policies so far. I decided to codify them so that others, should they choose to contribute, may follow them as well, and so that inspectors of the codebase may understand the purpose of the branches - In particular, why both `dev-master` and `master` exist, and the difference between them.