Add config parameter for play channels
This parameter should be an array of strings.

They are specified in priority order: That is, strings which appear earlier in the array are prioritized over those which appear later.

Afterwards, `play` shall choose channels with names according to the configured array, with all comparison done without case sensitivity:

First priority will be channels whose name matches exactly the string in the array.

Second priority will be channels whose name starts with the string in the array.

Last priority will be channels whose name includes, at any position, the string in the array. This may be done with a less flexible case insensitivity than the others: The others will use caselessCompare, which is locale-sensitive, while this shall simply use includes() on strings converted to uppercase.