Add playChannelSelector()
This function will do the work of choosing a channel out of all those available to `play`.

This implementation is not correct to the tiebreakers I specified in the commit description of c73761e59cd159a969bc9bbab8116af5029baadd and #22. While the union of those two specifies an ordering which is rigid and doesn't depend, for first ties, on ordering. Instead, the prescribed order uses information about the channel name to break ties when possible.

This is, however, somewhat more complicated to implement than this more naive implementation. Once I've got this working and on `dev-master`, I'll redo it a bit to match the strict requirements.

I'm also not certain if this processes in the correct order - That is, I'm not certain if the order in the array is the same as the position order of the channels (as given by .calculatedPosition). If this is not the case, then I'll have the function sort on calculated position (ascending) before iterating over the channels.