Add configurable custom commands
As an overly powerful solution to #29.

There are two types of custom commands: Those where the key is "equal to" the command, and the value is the exact output message, and those where the key is what the command "starts with", and the value is a format-string where "%s" represents the rest of the message.

These commands will not appear in the help message (they act like easter eggs). The standard commands will always take preference over the customs.