Add replyOnly commands
First, switch the please/please don't commands to repliers, as they do
work better that way.

Then, add an 'introduce to me' command, which extends an existing
command as a static string reply using the reply mechanism.

This maintains the principle that each 'sort' of custom command is
represented in default configuration.