Add event handling on the broadcast
Testing shows that the stream tends to end uncomfortably often
(Overnight, the stream ended four times, without any errors: Two of
these times happened within three seconds of each other). This occurs
without errors or warnings...

So it's important that CadenceBot reconnects on stream end, which was
done in testing and meant voice was still working even after multiple
disconnects in the morning.

The error and warning handlers are there just for logging: CadenceBot's
primary function is the stream, and I want to see any events that are
happening.

Additionally, this logging has the added benefit of acting as a watchdog
on Cadence mainline: If the stream encounters issues, it may generate
logging on CadenceBot as well, which has the potential to assist in
diagnosis.