Add and implement multitargeted custom commands
These are custom commands which can have an arbitrary number of mentions
(as %un, where n is a natural number). The format requires an expected
count of total mentions for verification purposes: All parseformats must
include that many mentions. This is not an engineering requirement - But
it forces, for verification, that the correct number of mentions are
given.

The same number of mentions are not required to appear in the output
string: Indeed, all that's required is that no %u mention appears in the
format string that did not appear in the parse string.