Implement recognition of time-windowed bans
Bans can now be objects instead of simple IDs - If a ban is an object,
the target ID should be in the 'ID' field.

The object can specify start and end times in any format parsable by
Date.parse, but does not need to - A ban will only be applied during the
time interval specified.

This option can be specified manually in configuration, but is not
intended to be, especially thanks to the specification of absolute
dates: It will be used soon for the Dynamic Ban admin-only feature.