Add config for two command aliases
This is a minimal set to provide examples for how one can write aliases,
as I am not certain if I want shorthand or lowercase aliases - But these
are safe. We are more or less agreeing across the board that
'nowplaying' and 'now playing' are both somewhat valid, and Kelvin had
some interesting misinterpretations that play could handle a song
request - I can see why this would be intuitive, so I decided to allow
it.