Fix disabling most customCommands directly
Jun noticed a logic error I made where I checked whether an alias or an
alias target was disabled, but not whether a directly referenced command
was disabled, for startsWith, targeted, and multitargeted commands.

I probably made this error reasoning that I had already checked disabled
once above, but did not consider that this branch had two requirements -
And thus not having an alias would be the same as not being disabled.